+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import { ThemeProvider } from '@emotion/react';
 import { Outlet } from 'react-router-dom';
-import { ModalProvider } from './features/ui/Modal/ModalContext';
+import { ModalProvider } from './components/@common/Modal/ModalContext';
 import { theme } from './styles/theme';
 
 const App = () => {
