+++ frontend/src/layouts/Layout.stories.tsx
@@ -2,8 +2,8 @@ import CloseIcon from '@/assets/close.svg';
 import BackButton from '@/components/@common/BackButton/BackButton';
 import Button from '@/components/@common/Button/Button';
 import { ColorKey } from '@/constants/color';
-import { ModalProvider } from '@/features/ui/Modal/ModalContext';
-import useModal from '@/features/ui/Modal/useModal';
+import { ModalProvider } from '@/components/@common/Modal/ModalContext';
+import useModal from '@/components/@common/Modal/useModal';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
 import { ReactNode } from 'react';
 import Layout from './Layout';
